add specific tests for gptorch
