+++ tests/models/test_cnp.py
